requirements.txt

numpy==1.21.2
pandas==1.3.2
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
requests==2.18.3
SQLAlchemy-redshift==0.8.5
